requirements.txt
streamlit
openai
python-docx
groq